+++ test/inertia_test.exs
@@ -12,6 +12,14 @@ defmodule InertiaTest do
     :ok
   end
 
+  test "checking if a response is Inertia-rendered", %{conn: conn} do
+    conn = get(conn, "/")
+    assert Inertia.Controller.inertia_response?(conn)
+
+    conn = conn |> recycle() |> get("/non_inertia")
+    refute Inertia.Controller.inertia_response?(conn)
+  end
+
   test "renders JSON response with x-inertia header", %{conn: conn} do
     conn =
       conn