+++ test/support/my_app/lib/my_app_web/controllers/page_html.ex
@@ -1,0 +1,8 @@
+defmodule MyAppWeb.PageHTML do
+  use MyAppWeb, :html
+
+  def non_inertia(assigns) do
+    ~H"""
+    """
+  end
+end